update working service/pcswap

+++ src/main/java/com/example/demo/repository/MultipliersRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.model.Multipliers;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MultipliersRepository extends JpaRepository<Multipliers, Long> {
-    Multipliers findMultipliersById(int id);
+    Multipliers findMultipliersById(long id);
 }
